% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tools.R
\name{build_and_list}
\alias{build_and_list}
\title{Build, Install, and List Tutorials}
\usage{
build_and_list(package = "RMastery")
}
\arguments{
\item{package}{Package name (default: "RMastery")}
}
\description{
Automates documentation, installation, and listing tutorials
}
