% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tutorials.R
\name{list_tutorials}
\alias{list_tutorials}
\title{List Available Tutorials}
\usage{
list_tutorials(package = "RMastery", print = TRUE)
}
\arguments{
\item{package}{Name of the package (default: "RMastery")}

\item{print}{Logical; whether to print nicely formatted output}
}
\value{
Data frame of tutorials
}
\description{
Returns a data frame of all available learnr tutorials in the package.
}
