% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tutorial.R
\name{load_tutorial}
\alias{load_tutorial}
\title{Load a Given Tutorial from the Package}
\usage{
load_tutorial(package = "RMastery", tutorial)
}
\arguments{
\item{package}{Package name (default: "RMastery")}

\item{tutorial}{Tutorial name (character)}
}
\description{
Load a Given Tutorial from the Package
}
