% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_learning.R
\name{start_learning}
\alias{start_learning}
\title{Start the RMastery Learning Experience}
\usage{
start_learning()
}
\description{
This function launches an interactive menu for users to select and run tutorials.
}
